requirements.txt
fastapi
uvicorn[standard]
requests
PyMuPDF
sentence-transformers
redis
python-dotenv
pydantic
numpy
faiss-cpu==1.12.0
pdfplumber
langchain
langchain-community
langchain-openai
openai
httpx
python-multipart
fuzzywuzzy
python-levenshtein
langdetect
google-generativeai
aiohttp
